No Clocks pkgdown

Build a pkgdown::pkgdown() site with additional options to include extra reports.

The reports that can be included are:

Test Results Report via testdown::test_down()

Test Coverage Results via covrpage::covrpage()

Git Reports via gitdown::git_down()

Usage

build_pkgdown_with_reports(
  pkg = getwd(),
  pkgdown_path = "public",
  assets_path = "pkgdown/assets",
  reports = c("coverage", "testdown", "gitdown"),
  git_branch_ref = "main",
  overwrite_assets = TRUE
)

Arguments

pkg

(character) Path to the package in development

pkgdown_path

(character) Relative path inside the package to store the final pkgdown site

assets_path

(character) Relative path within the package to store the pkgdown assets

reports

(character) Vector of reports to be produced. Must be a subset of c("testdown","gitdown", "coverage"). The default is c("coverage", "testdown", "gitdown").

git_branch_ref

(character) Git branch to use. `main' by default

overwrite_assets

(logical) Whether the assets directory should be overwritten. Default is TRUE.

Value

None Generate a pkgdown with test and coverage reports

Examples

if (FALSE) { # \dontrun{
# build_pkgdown_with_reports(
#    pkg = ".",
#    pkgdown_path = "public",
#    assets_path = "pkgdown/assets",
#    reports = c("testdown","coverage")
} # }